EDS-73 Auto-follow cookie.  Set the cookie
https://issues.shibboleth.net/jira/browse/EDS-73

Add code to manage the "Auto Follow" cookie.  On clicking one of the radio
buttons the "auto-follow" cookie is immediately set or cleared.

The number of buttons is configurable (but the config changes have to be reflected in the languages)

All text is Globalized and messages put in for US-EN.

New CSS for the selector.